+++ tests/integration/targets/param_list_compare/tasks/include/simple.yaml
@@ -1,6 +1,6 @@
 ---
 - debug:
-    msg: "START resource_list_compare integration tests on connection={{ ansible_connection }}"
+    msg: "START param_list_compare integration tests on connection={{ ansible_connection }}"
 
 - name: Setup supported resource module list json
   ansible.builtin.set_fact:
@@ -45,7 +45,7 @@
 
 - name: Get the final list of resources
   set_fact:
-    final_network_resources: "{{ network_resources['modules']|ansible.utils.resource_list_compare(provided_resources) }}"
+    final_network_resources: "{{ network_resources['modules']|ansible.utils.param_list_compare(provided_resources) }}"
 
 - name: Assert final network resources
   assert:
@@ -53,4 +53,4 @@
       - "{{ expected_network_resources | symmetric_difference(final_network_resources['actionable']) |length\
               \ == 0 }}"
 - debug:
-    msg: "END resource_list_compare integration tests on connection={{ ansible_connection }}"
+    msg: "END param_list_compare integration tests on connection={{ ansible_connection }}"
