+++ tests/unit/plugins/filter/test_param_list_compare.py
@@ -10,12 +10,12 @@ __metaclass__ = type
 
 import unittest
 from ansible.errors import AnsibleFilterError
-from ansible_collections.ansible.utils.plugins.filter.resource_list_compare import (
-    resource_list_compare,
+from ansible_collections.ansible.utils.plugins.filter.param_list_compare import (
+    param_list_compare,
 )
 
 
-class TestResource_list_compare_merge(unittest.TestCase):
+class TestParam_list_compare_merge(unittest.TestCase):
     def test_valid_data(self):
         """Check passing valid data as per criteria"""
 
@@ -23,7 +23,7 @@ class TestResource_list_compare_merge(unittest.TestCase):
         other = ["all"]
         args = [base, other]
         kwargs = {}
-        result = resource_list_compare(*args, **kwargs)
+        result = param_list_compare(*args, **kwargs)
         self.assertEqual(result["actionable"], base)
 
     def test_valid_data_with_not_bang(self):
@@ -34,7 +34,7 @@ class TestResource_list_compare_merge(unittest.TestCase):
         args = [base, other]
         expected = ["interfaces", "l3_interfaces"]
         kwargs = {}
-        result = resource_list_compare(*args, **kwargs)
+        result = param_list_compare(*args, **kwargs)
         self.assertEqual(result["actionable"], expected)
 
     def test_invalid_args_length_data(self):
@@ -44,7 +44,7 @@ class TestResource_list_compare_merge(unittest.TestCase):
         args = [base]
         kwargs = {}
         with self.assertRaises(AnsibleFilterError) as error:
-            resource_list_compare(*args, **kwargs)
+            param_list_compare(*args, **kwargs)
         self.assertIn(
             "Missing either 'base' or 'other value in filter input",
             str(error.exception),
@@ -58,7 +58,7 @@ class TestResource_list_compare_merge(unittest.TestCase):
         args = [base, other]
         kwargs = {}
         with self.assertRaises(AnsibleFilterError) as error:
-            resource_list_compare(*args, **kwargs)
+            param_list_compare(*args, **kwargs)
         self.assertIn("cannot be converted to a list", str(error.exception))
 
     def test_invalid_other_type_data(self):
@@ -69,7 +69,7 @@ class TestResource_list_compare_merge(unittest.TestCase):
         args = [base, other]
         kwargs = {}
         with self.assertRaises(AnsibleFilterError) as error:
-            resource_list_compare(*args, **kwargs)
+            param_list_compare(*args, **kwargs)
         self.assertIn("cannot be converted to a list", str(error.exception))
 
     def test_invalid_unsupported_bang(self):
@@ -80,5 +80,5 @@ class TestResource_list_compare_merge(unittest.TestCase):
         args = [base, other]
         kwargs = {}
         with self.assertRaises(AnsibleFilterError) as error:
-            resource_list_compare(*args, **kwargs)
+            param_list_compare(*args, **kwargs)
         self.assertIn("The following are unsupported", str(error.exception))
